Rework battle engine

+++ src/main/java/com/github/retro_game/retro_game/dto/CombatReportUnitGroupDto.java
@@ -1,19 +1,19 @@
 package com.github.retro_game.retro_game.dto;
 
 public class CombatReportUnitGroupDto {
-  private final int numUnits;
+  private final long numUnits;
   private final double weapons;
   private final double shields;
   private final double armor;
 
-  public CombatReportUnitGroupDto(int numUnits, double weapons, double shields, double armor) {
+  public CombatReportUnitGroupDto(long numUnits, double weapons, double shields, double armor) {
     this.numUnits = numUnits;
     this.weapons = weapons;
     this.shields = shields;
     this.armor = armor;
   }
 
-  public int getNumUnits() {
+  public long getNumUnits() {
     return numUnits;
   }
